Update for chatty API changes
The Chatty API has changes so that you can no longer request a full day's worth of posts.  So there is now a second program that currently I run 3 times a day to collect the current post IDs on the front page (18 hours), and write the post number of date / time of post to a file.  This file is then used by the Post Reporter to remember which post (root thread ids) where made each day.  The thread ids are kept for 40 days (we only need a months worth at most, when doing a monthly report).
Various changes and fixes.
New bugs (most popular emoji list, for one, is all on one line)